@@ -15,24 +15,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.3",
-    "@angular/common": "~13.0.3",
-    "@angular/compiler": "~13.0.3",
-    "@angular/core": "~13.0.3",
-    "@angular/forms": "~13.0.3",
-    "@angular/platform-browser": "~13.0.3",
-    "@angular/platform-browser-dynamic": "~13.0.3",
-    "@angular/router": "~13.0.3",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "rxjs": "^7.4.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.4",
-    "@angular/cli": "^13.0.4",
-    "@angular/compiler-cli": "~13.0.3",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/jquery": "^3.5.5",
     "@types/node": "^16.11.11",
-    "typescript": "~4.4.4"
+    "typescript": "~4.8.4"
   }
 }
